+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
 
 
-## [1.1.0](https://github.com/futurestudio/hapi-rate-limitor/compare/v1.0.0...v1.1.0) - 2019-11-xx
+## [2.0.0](https://github.com/futurestudio/hapi-jwt/compare/v1.1.0...v2.0.0) - 2020-xx-xx
+
+### Updated
+- bump dependencies
+- switch packages from `jws` to `jose`
+
+
+### Breaking Changes
+- requires Node.js v12
+
+
+## [1.1.0](https://github.com/futurestudio/hapi-jwt/compare/v1.0.0...v1.1.0) - 2019-11-xx
 
 ### Added
 - checks to ensure a token when decoding a JWT